# add serializing numpy boolean