--- a/src/numbuf/python/src/pynumbuf/adapters/numpy.cc
+++ b/src/numbuf/python/src/pynumbuf/adapters/numpy.cc
@@ -32,6 +32,7 @@ Status SerializeArray(PyArrayObject* array, SequenceBuilder& builder,
     std::vector<PyObject*>& subdicts, std::vector<PyObject*>& tensors_out) {
   int dtype = PyArray_TYPE(array);
   switch (dtype) {
+    case NPY_BOOL:
     case NPY_UINT8:
     case NPY_INT8:
     case NPY_UINT16: